chore: Move tests into dedicated file
